# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Competition Management API",
        "version": "1.0.0",
        "description": "API documentation for Competition Management System."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        },
        {
            "url": "https://node-test-fu37.onrender.com",
            "description": "Deployment server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter 'Bearer' followed by a space and the token. Example: 'Bearer abc123'"
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/competitions": {
            "get": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Get all competitions",
                "description": "Retrieve a list of all competitions including the count of regions.",
                "responses": {
                    "200": {
                        "description": "A list of competitions.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Competition"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
...
        "/api/competitions/{id}": {
            "put": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Update a competition",
                "description": "Update details of an existing competition.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "integer",
                        "description": "ID of the competition to update."
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateCompetition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Competition updated successfully."
                    },
                    "400": {
                        "description": "Validation errors."
                    },
                    "403": {
                        "description": "Access denied. Admin rights required."
                    },
                    "404": {
                        "description": "Competition not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Delete a competition",
                "description": "Delete a competition by its ID.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "integer",
                        "description": "ID of the competition to delete."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Competition deleted successfully."
                    },
                    "403": {
                        "description": "Access denied. Admin rights required."
                    },
                    "404": {
                        "description": "Competition not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Competition": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "venue": {
                    "type": "string"
                },
                "registration_deadline": {
                    "type": "string",
                    "format": "date"
                },
                "maximum_teams": {
                    "type": "integer"
                },
                "fees": {
                    "type": "number",
                    "format": "float"
                },
                "rules": {
                    "type": "string"
                },
                "pdf_url": {
                    "type": "string"
                },
                "regions_count": {
                    "type": "integer"
                }
            }
        },
...
    }
}